Make generated struct be !Send & !Sync & !Unpin
This is done via a combination of `PhantomData`, `*mut u8`, and
`PhantomPinned`.

Re-exported `core` to prevent failure if `core` refers to another module
at the call site.

Also moved `*mut u8` out of the empty array and into `PhantomData`,
because having it in the array makes the generated struct require
alignment of `*mut u8`. This is dangerous because `&T` must always be
properly aligned.

Corrected the static assertion to `assert_not_impl_any` to ensure that
the generated type does not implement any of these autos traits.